+++ backend/src/main/java/com/spotchzxk/shared/exception/InsufficientFollowerCountException.java
@@ -1,7 +1,7 @@
 package com.spotchzxk.shared.exception;
 
 public class InsufficientFollowerCountException extends RuntimeException {
-    public InsufficientFollowerCountException(String channelId, int followerCount) {
+    public InsufficientFollowerCountException(String channelId, long followerCount) {
         super("팔로워 100명 미만의 채널은 등록할 수 없습니다: " + channelId + " (현재 " + followerCount + "명)");
     }
 }
